#368: Restored checking for export-client start and other fixes.
