+++ internal/export/registration.go
@@ -81,14 +81,6 @@ type Filter struct {
 	ValueDescriptorIDs []string `bson:"valueDescriptorIdentifiers,omitempty" json:"valueDescriptorIdentifiers,omitempty"`
 }
 
-// EncryptionDetails - Provides details for encryption
-// of export data per client request
-type EncryptionDetails struct {
-	Algo       string `bson:"encryptionAlgorithm,omitempty" json:"encryptionAlgorithm,omitempty"`
-	Key        string `bson:"encryptionKey,omitempty" json:"encryptionKey,omitempty"`
-	InitVector string `bson:"initializingVector,omitempty" json:"initializingVector,omitempty"`
-}
-
 func (reg Registration) Validate() (bool, error) {
 
 	if reg.Name == "" {
